cambios

+++ trianatourist/src/main/java/com/salesianostriana/dam/trianatourist/servicios/CategoryService.java
@@ -3,7 +3,9 @@ package com.salesianostriana.dam.trianatourist.servicios;
 import com.salesianostriana.dam.trianatourist.errores.excepciones.ListEntityNotFoundException;
 import com.salesianostriana.dam.trianatourist.errores.excepciones.SingleEntityNotFoundException;
 import com.salesianostriana.dam.trianatourist.modelos.Category;
+import com.salesianostriana.dam.trianatourist.modelos.POI;
 import com.salesianostriana.dam.trianatourist.repositorios.CategoryRepository;
+import com.salesianostriana.dam.trianatourist.repositorios.POIRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import java.util.List;
@@ -14,6 +16,7 @@ import java.util.UUID;
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final POIRepository poiRepository;
 
     public List<Category> findAll() {
         List<Category> lista = categoryRepository.findAll();
@@ -38,7 +41,16 @@ public class CategoryService {
     }
 
     public void deleteById (UUID id){
-        categoryRepository.deleteById(id);
+        if(categoryRepository.findById(id).isEmpty()){
+            throw new SingleEntityNotFoundException(id.toString(), Category.class);
+        }else{
+            List<POI>pois = poiRepository.findByCategoryId(id);
+            for (POI p: pois) {
+                p.setCategory(null);
+            }
+            categoryRepository.deleteById(id);
+        }
+
     }
 
    // public Category findByName(String name){ return categoryRepository.findByName(name); }
